Refactor
Removes the nav-close component and its redux.
I created a redux toggleNav that receives by parameter the status of the menu.
The status of the last menu is always the opposite of the current one, for example from true to false.
In the nav-icon-menu component and I put the effect of when clicking the menu it turns an x ​​so the user knows visually that it needs to click on it to close the menu.
Set the test of icon-nav-mobile to not give errors.